requirements.txt
botbuilder-core==4.14.8
botbuilder-integration-aiohttp==4.14.8
botbuilder-schema==4.14.8
python-jenkins==1.8.0
requests==2.31.0
python-dotenv==1.0.1
aiosqlite==0.20.0

aiohttp~=3.8.5